add test for empty app & refactor a bit
